backend/pyproject.toml
[project]
name = "donos-backend"
version = "0.1.0"
description = "Donos API — transparent donation platform on XRPL"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn>=0.42.0",
    "xrpl-py>=4.5.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.0",
    "httpx>=0.28.1",
    "pytest>=8.4.0",
    "supabase>=2.28.2",
]

[project.scripts]
dev = "app.main:run"
ops = "app.cli:main"
